pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Prithiviraj25",email = "prithivirajn25@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "langchain (>=1.2.15,<2.0.0)",
    "langchain-openai (>=1.1.13,<2.0.0)",
    "langgraph (>=1.1.6,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
